requirements.txt
apache_beam[gcp]
google-cloud-pubsub
typing
termcolor
